TrinityBackendFastAPI/app/features/select_models_feature_based/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
motor==3.3.2
minio==7.2.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
